refactor and pr feedback
